requirements.txt
aiogram==3.7.0
openai>=1.30.0
pydub==0.25.1
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-core==2.16.1
aiohttp==3.9.3
python-pptx==0.6.23